LPD-62099 Move comment service
